+++ crates/socketioxide-mongodb/CHANGELOG.md
@@ -1,2 +1,12 @@
+# socketioxide-mongodb 0.1.3
+* fix: rustdoc failed build
+
+# socketioxide-mongodb 0.1.2
+* fix: race condition between ack timeout and adapter request timeout when broadcasting with acks.
+
+# socketioxide-mongodb 0.1.1
+* fix: compilation error without default features.
+* chore(deps): bump bson to 3.0.0
+
 # socketioxide-mongodb 0.1.0
 * Initial release!